introduced 2 type evaluation
this is to prevent the previous issues of opponent being 1 step off from AI

        # doesn't terminate properly
        # for example
        # 2 depth on _OO_O_
        # depth 2: _OOOO_
        # depth 1: _OOOOX <- this is worthless
        # so O will look for something else at depth 2
        # but in fact, _OOOOX should worth a lot to O

I am fixing this for odd depth, so in this case type1 eval is basically 1 step behind type 2 eval.

Type1 eval is used for the current player (AI). Type2 eval is used for the opponent (human).